workflow: move tracker kind helper out of loader
